requirements.txt
openai
python-dotenv
flask
markdown
pymongo